+++ app/javascript/application.js
@@ -1,6 +1,6 @@
 // Configure your import map in config/importmap.rb. Read more: https://github.com/rails/importmap-rails
-import "@hotwired/turbo-rails"
-import "controllers"
-import "chartkick"
-import "Chart.bundle"
-import "running"
+import "@hotwired/turbo-rails";
+import "controllers";
+import "chartkick";
+import "Chart.bundle";
+import "runningcircle";
